Revert "Merge pull request #291 from moonboots/simplify-calendar-day-dom"
This reverts commit 629d8c0b9ac6365299483b0fcac5917d2788ffa0, reversing
changes made to 64785993ee2ff8511c057d3d49e88b3813d54c0b.
