+++ src/components/CalendarMonth.jsx
@@ -48,6 +48,10 @@ const defaultProps = {
   monthFormat: 'MMMM YYYY', // english locale
 };
 
+export function getModifiersForDay(modifiers, day) {
+  return day ? Object.keys(modifiers).filter(key => modifiers[key](day)) : [];
+}
+
 export default class CalendarMonth extends React.Component {
   constructor(props) {
     super(props);
@@ -101,18 +105,26 @@ export default class CalendarMonth extends React.Component {
           <tbody className="js-CalendarMonth__grid">
             {weeks.map((week, i) => (
               <tr key={i}>
-                {week.map((day, dayOfWeek) => (
-                  <CalendarDay
-                    day={day}
-                    isOutsideDay={!day || day.month() !== month.month()}
-                    modifiers={modifiers}
-                    key={dayOfWeek}
-                    onDayMouseEnter={onDayMouseEnter}
-                    onDayMouseLeave={onDayMouseLeave}
-                    onDayClick={onDayClick}
-                    renderDay={renderDay}
-                  />
-                ))}
+                {week.map((day, dayOfWeek) => {
+                  const modifiersForDay = getModifiersForDay(modifiers, day);
+                  const className = cx('CalendarMonth__day', {
+                    'CalendarMonth__day--outside': !day || day.month() !== month.month(),
+                  }, modifiersForDay.map(mod => `CalendarMonth__day--${mod}`));
+
+                  return (
+                    <td className={className} key={dayOfWeek}>
+                      {day &&
+                        <CalendarDay
+                          day={day}
+                          onDayMouseEnter={onDayMouseEnter}
+                          onDayMouseLeave={onDayMouseLeave}
+                          onDayClick={onDayClick}
+                          renderDay={renderDay}
+                        />
+                      }
+                    </td>
+                  );
+                })}
               </tr>
             ))}
           </tbody>
